requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.2
daphne==4.0.0
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-standardized-errors==0.12.5
executing==1.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
msgpack==1.0.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
wcwidth==0.2.6
zope.interface==6.0
